We're calling them "plugins" now
The WIP CLI tool is named `sopel-plugins`. We'll likely add support for
loading entry points under the `sopel.plugins` group. So on, so forth.

Repo has been renamed on GitHub already, so this is just updating the
prototype branch to match it.

Not 100% sure what the subcommand will be to show a plugin's details (or
even if we'll have one), so I'm not touching the `sopel-plugins show`
conceit yet. It's still good enough for a prototype.